Get everything working from previous commit Move the model code out of frame.js/view.js and into model.js rename clientglue to controller Fix imports given new relative location of files Remove references to the Profile Comment out code related to Save Full Page Move the handler for GET /:id/:domain out of root server.js and into page-specific server.js Part of the work for #1713
